requirements.txt
discord.py==1.5.1
python-decouple==3.3

